+++ tests/linksAndImages.test.js
@@ -3,6 +3,12 @@ import twitchdown from '../index';
 
 test('parses links', (t) => {
   t.deepEqual([ { type: 'a', props: { key: 0, href: 'http://example.com' }, children: [ 'World' ] } ], twitchdown('[World](http://example.com)'));
+  t.deepEqual([ 'hello ', { type: 'a', props: { key: 0, href: 'http://example.com' }, children: [ 'World' ] }, ' ago' ], twitchdown('hello [World](http://example.com) ago'));
+  t.deepEqual([ 'hello', { type: 'a', props: { key: 0, href: 'http://example.com' }, children: [ 'World' ] }, 'ago' ], twitchdown('hello[World](http://example.com)ago'));
+});
+
+test('newline before and after link rendered as space', (t) => {
+  t.deepEqual([ 'hello ', { type: 'a', props: { key: 0, href: 'http://example.com' }, children: [ 'World' ] }, ' again' ], twitchdown('hello\n[World](http://example.com)\nagain'));
 });
 
 test('parses anchor links', (t) => {
@@ -41,3 +47,8 @@ test('should not add href to link if do not have href or ref', (t) => {
   t.deepEqual([ { type: 'a', props: { key: 0, }, children: [ 'World' ] } ], twitchdown('[World]'));
 });
 
+test('sets external links to open in a new window (_blank)', (t) => {
+  t.deepEqual([ { type: 'a', props: { key: 0, href: 'http://example.com', target: '_blank' }, children: [ 'World' ] } ], twitchdown('[World](http://example.com)', { openExternalInNewWindow: true }));
+  t.deepEqual([ { type: 'a', props: { key: 0, href: '#example' }, children: [ 'Example' ] } ], twitchdown('[Example](#example)', { openExternalInNewWindow: true }));
+  t.deepEqual([ 'hello ', { type: 'a', props: { key: 0, href: 'http://example.com', target: '_blank' }, children: [ 'World' ] }, '!' ], twitchdown('\nhello [World]!', { openExternalInNewWindow: true, referenceLinks: { world: 'http://example.com' } }));
+});