+++ CHANGELOG.md
@@ -4,9 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [Unreleased]
 
-## [1.3.3] - 2019-07-29
+## [1.4.0] - 2020-02-18
+
+## Added
+- `openExternalInNewWindow` option to open external links in a new window
+
+## Fixed
+- Handling of sub-lists and multi-line items in lists
+
+## [1.3.3] - 2020-02-15
 
 ## Changed
 - Fixed handling of p tags when using paragraphs option
@@ -75,6 +83,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## 1.0.0 - 2018-07-10
 Initial Release
 
+[unreleased]: https://gitlab.com/bytesnz/twitchdown/compare/v1.4.0...dev
+[1.4.0]: https://gitlab.com/bytesnz/twitchdown/compare/v1.3.3...v1.4.0
 [1.3.3]: https://gitlab.com/bytesnz/twitchdown/compare/v1.3.2...v1.3.3
 [1.3.2]: https://gitlab.com/bytesnz/twitchdown/compare/v1.3.1...v1.3.2
 [1.3.1]: https://gitlab.com/bytesnz/twitchdown/compare/v1.3.0...v1.3.1